requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking requirements.in
#
affine==2.4.0
    # via rasterio
aiobotocore==2.5.0
    # via s3fs
aiohttp==3.8.4
    # via
    #   aiobotocore
    #   gcsfs
    #   s3fs
aioitertools==0.11.0
    # via aiobotocore
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via
    #   aiohttp
    #   fiona
    #   rasterio
botocore==1.29.76
    # via aiobotocore
cachetools==5.3.1
    # via google-auth
cdsapi==0.6.1
    # via -r requirements.in
certifi==2023.5.7
    # via
    #   fiona
    #   netcdf4
    #   pyproj
    #   rasterio
    #   requests
cftime==1.6.2
    # via netcdf4
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   click-plugins
    #   cligj
    #   fiona
    #   openhexa-sdk
    #   rasterio
click-plugins==1.1.1
    # via
    #   fiona
    #   rasterio
cligj==0.7.2
    # via
    #   fiona
    #   rasterio
decorator==5.1.1
    # via gcsfs
dill==0.3.6
    # via multiprocess
fiona==1.9.4.post1
    # via geopandas
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.6.0
    # via
    #   -r requirements.in
    #   gcsfs
    #   s3fs
gcsfs==2023.6.0
    # via -r requirements.in
geopandas==0.13.2
    # via -r requirements.in
google-api-core==2.11.1
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.20.0
    # via
    #   gcsfs
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
google-auth-oauthlib==1.0.0
    # via gcsfs
google-cloud-core==2.3.2
    # via google-cloud-storage
google-cloud-storage==2.9.0
    # via gcsfs
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.5.0
    # via google-cloud-storage
googleapis-common-protos==1.59.1
    # via google-api-core
idna==3.4
    # via
    #   requests
    #   yarl
importlib-metadata==6.7.0
    # via fiona
jmespath==1.0.1
    # via botocore
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.14
    # via openhexa-sdk
netcdf4==1.6.4
    # via -r requirements.in
numpy==1.25.0
    # via
    #   -r requirements.in
    #   cftime
    #   netcdf4
    #   pandas
    #   pyarrow
    #   rasterio
    #   shapely
    #   snuggs
    #   xarray
oauthlib==3.2.2
    # via requests-oauthlib
openhexa-sdk==0.1.21
    # via -r requirements.in
packaging==23.1
    # via
    #   geopandas
    #   xarray
pandas==2.0.2
    # via
    #   -r requirements.in
    #   geopandas
    #   xarray
protobuf==4.23.3
    # via
    #   google-api-core
    #   googleapis-common-protos
pyarrow==12.0.1
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pyparsing==3.1.0
    # via snuggs
pyproj==3.6.0
    # via geopandas
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   botocore
    #   pandas
pytz==2023.3
    # via pandas
pyyaml==6.0
    # via openhexa-sdk
rasterio==1.3.7
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   cdsapi
    #   gcsfs
    #   google-api-core
    #   google-cloud-storage
    #   openhexa-sdk
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
s3fs==2023.6.0
    # via -r requirements.in
shapely==2.0.1
    # via geopandas
six==1.16.0
    # via
    #   fiona
    #   google-auth
    #   python-dateutil
snuggs==1.4.7
    # via rasterio
stringcase==1.2.0
    # via openhexa-sdk
tqdm==4.65.0
    # via cdsapi
typing-extensions==4.6.3
    # via aioitertools
tzdata==2023.3
    # via pandas
urllib3==1.26.16
    # via
    #   botocore
    #   google-auth
    #   requests
wrapt==1.15.0
    # via aiobotocore
xarray==2023.5.0
    # via -r requirements.in
yarl==1.9.2
    # via aiohttp
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
